fixed some bugs
their is a bug in notationToIndex. the bug is that suppose Rxd7 is a move but two rooks lets say on d8 and f7 can capture that d7 piece but the problem is f7 rook is pinned by queen or bishop on b3 so instead of saying Rdxd7 of R8xd7 lichess says Rxd7 as only d8 rook can capture the d7 piece. this caused some bugs in the program which i am currently solving. but other than that i also solved some other bugs occuring due to enpassent.